+++ eras/allegra/impl/src/Cardano/Ledger/Allegra/Core.hs
@@ -1,15 +1,10 @@
-{-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE UndecidableSuperClasses #-}
-
 module Cardano.Ledger.Allegra.Core (
   AllegraEraTxBody (..),
+  ValidityInterval (..),
   module Cardano.Ledger.Shelley.Core,
 )
 where
 
 import Cardano.Ledger.Allegra.Scripts (ValidityInterval (..))
+import Cardano.Ledger.Allegra.TxBody (AllegraEraTxBody (..))
 import Cardano.Ledger.Shelley.Core
-import Lens.Micro (Lens')
-
-class ShelleyEraTxBody era => AllegraEraTxBody era where
-  vldtTxBodyL :: Lens' (TxBody era) ValidityInterval