+++ eras/conway/impl/src/Cardano/Ledger/Conway/TxBody.hs
@@ -51,7 +51,6 @@ module Cardano.Ledger.Conway.TxBody (
   conwayProposalsDeposits,
 ) where
 
-import Cardano.Ledger.Allegra.Scripts (ValidityInterval (..))
 import Cardano.Ledger.Alonzo.TxAuxData (AuxiliaryDataHash (..))
 import Cardano.Ledger.Babbage.TxBody (
   BabbageTxBody (..),